backend/requirements.txt
Django==4.2.21
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]==3.2.13
